third_party/blink/web_tests/fast/dom/MutationObserver/mutation-record-constructor-expected.txt

MutationRecord should be exposed on window but not constructable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.MutationRecord is non-null.
PASS typeof MutationRecord is "function"
PASS new MutationRecord threw exception TypeError: Failed to construct 'MutationRecord': Illegal constructor.
PASS record instanceof MutationRecord is true
PASS successfullyParsed is true

TEST COMPLETE
